# packages/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Nexus API
  description: Local-first AI document analysis and knowledge graph platform
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: local development environment
paths:
...
  /workspaces/{workspaceId}:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get workspace by ID
      tags: [workspaces]
      operationId: getWorkspace
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update workspace
      tags: [workspaces]
      operationId: updateWorkspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete workspace
      tags: [workspaces]
      operationId: deleteWorkspace
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/files:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: List files in workspace
      tags: [files]
      operationId: listFiles
      parameters:
        - name: directoryId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Filter by directory ID
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/files/upload:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    post:
      summary: Upload a file
      tags: [files]
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: The file to upload
                directoryId:
                  type: string
                  format: uuid
                  nullable: true
                  description: Target directory (optional)
                tags:
                  type: array
                  items:
                    type: string
                  description: File tags (optional)
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/files/{fileId}:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: fileId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Get file metadata
      tags: [files]
      operationId: getFile
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadata'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete a file
      tags: [files]
      operationId: deleteFile
      responses:
        '204':
          description: No Content
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/files/{fileId}/download:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: fileId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      summary: Download file content
      tags: [files]
      operationId: downloadFile
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/documents/{documentId}/process:
    parameters: 
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Start document processing
      operationId: processDocument
      tags:
        - documents
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                chunk_size:
                  type: integer
                  default: 500
                  minimum: 100
                  maximum: 2000
                chunk_overlap:
                  type: integer
                  default: 50
                  minimum: 0
                  maximum: 500
                force_reprocess:
                  type: boolean
                  default: false
      responses:
        '202':
          description: Processing started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [processing]
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Document is already being processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /workspaces/{workspaceId}/documents/{documentId}/status:
    parameters: 
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get document processing status
      operationId: getDocumentStatus
      tags:
        - documents
      responses:
        '200':
          description: Processing status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [uploaded, processing, processed, failed]
                  progress:
                    type: object
                    nullable: true
                    properties:
                      current_step: 
                        type: string
                      percentage: 
                        type: integer
                      chunks_created: 
                        type: integer
                  error:
                    type: object
                    nullable: true
                    properties:
                      message:
                        type: string
                      code:
                        type: string
                  processed_at:
                    type: string
                    format: date-time
                    nullable: true
        '404': 
          $ref: '#/components/responses/NotFound'
        '500': 
          $ref: '#/components/responses/InternalServerError'
  /workspaces/{workspaceId}/documents/{documentId}/chunks:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: documentId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        schema:
          type: integer
          default: 1
          minimum: 1
      - name: limit
        in: query
        schema:
          type: integer
          default: 20
          minimum: 1
          maximum: 100
    get:
      summary: Get document chunks
      operationId: getDocumentChunks
      tags:
        - documents
      responses:
        '200':
          description: List of chunks
          content:
            application/json:
              schema:
                type: object
                properties:
                  chunks:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        chunk_index:
                          type: integer
                        content:
                          type: string
                        created_at:
                          type: string
                          format: date-time
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
        '404':
          $ref: '#/components/responses/NotFound'
  /workspaces/{workspaceId}/search:
    parameters:
      - name: workspaceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: RAG search
      operationId: searchWorkspace
      tags:
        - search
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties: 
                query:
                  type: string
                  example: "機械学習について教えて"
                  minLength: 1
                  maxLength: 1000
                top_k:
                  type: integer
                  default: 5
                  minimum: 1
                  maximum: 20
      responses:
        '200':
          description: Search results with generated answer
          content:
            application/json:
              schema:
                type: object
                required: [answer, sources]
                properties:
                  answer:
                    type: string
                    description: AI-generated answer based on retrieved context
                  sources:
                    type: array
                    description: Source chunks used for answer generation
                    items:
                      type: object
                      required: [document_id, chunk_index, content, score]
                      properties:
                        document_id:
                          type: string
                          format: uuid
                        chunk_index:
                          type: integer
                        content: 
                          type: string
                        score:
                          type: number
                          format: float
                          description: Similarity score from vector search
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    Workspace:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        settings:
          type: object
          nullable: true
          additionalProperties: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    FileMetadata:
      type: object
      required: [id, fileName, mimeType, sizeBytes, createdAt]
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
          description: Original filename
        mimeType:
          type: string
          description: MIME type (e.g., application/pdf)
        sizeBytes:
          type: integer
          format: int64
          description: File size in bytes
        sha256Hash:
          type: string
          description: SHA256 hash of file content
        createdAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - uploaded
            - processing
            - processed
            - failed
          description: Document processing status
        tags:
          type: array
          items:
            type: string
          nullable: true
          description: File tags for categorization
    FileListResponse:
      type: object
      required: [files, total]
      properties:
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileMetadata'
        total:
          type: integer
          description: Total number of files matching the query
    FileUploadResponse:
      type: object
      required: [id, fileName, mimeType, sizeBytes, createdAt]
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        mimeType:
          type: string
        sizeBytes:
          type: integer
          format: int64
        sha256Hash:
          type: string
          description: SHA256 hash for deduplication
        createdAt:
          type: string
          format: date-time
    
    # ========================================
    # Chat Schemas
    # ========================================
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: 
              type: string
            message: 
              type: string
            details: 
              type: object  
            timestamp: 
              type: string
            request_id: 
              type: string  
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'